+++ runtime/src/main/scala/com/ing/baker/runtime/actor/serialization/BakerProtobufSerializer.scala
@@ -39,7 +39,7 @@ class BakerProtobufSerializer(system: ExtendedActorSystem) extends SerializerWit
     Entry("baker.types.Type", classOf[com.ing.baker.types.Type], actor.protobuf.Type),
     Entry("baker.types.Value", classOf[com.ing.baker.types.Value], actor.protobuf.Value),
 
-    Entry("ProcessIndex.GetShardIndex", classOf[com.ing.baker.runtime.actor.ClusterBakerActorProvider.GetShardIndex], actor.process_index.protobuf.GetShardIndex),
+    Entry("ProcessIndex.GetShardIndex", classOf[com.ing.baker.runtime.actor.ClusterBakerActorApi.GetShardIndex], actor.process_index.protobuf.GetShardIndex),
 
     Entry("ProcessIndex.ActorCreated", classOf[ProcessIndex.ActorCreated], actor.process_index.protobuf.ActorCreated),
     Entry("ProcessIndex.ActorPassivated", classOf[ProcessIndex.ActorPassivated], actor.process_index.protobuf.ActorPassivated),